chore(panel): rename proposalClients store to snake_case
Content stores use snake_case filenames by convention; proposalClients.js
was the lone camelCase exception. File and test renamed via git mv, the
two source imports and two jest.mock paths updated. The Pinia store id
stays 'proposalClients' (renaming it would touch devtools/persistence
keys for no benefit).
