+++ frontend/stores/proposal_clients.js
